+++ src/main/java/co/istad/rentiq_api/features/category/cateogryDto/CategoryResponse.java
@@ -1,11 +1,13 @@
 package co.istad.rentiq_api.features.category.cateogryDto;
 
+import java.math.BigDecimal;
+
 public record CategoryResponse(
         Integer id,
         Integer parentId,
         String name,
         String slug,
-        Double commissionRate,
+        BigDecimal commissionRate,
         String iconUrl,
         Boolean active
 ) {
